Fix PVS implicit enum cast warning

+++ src/capture/capture_video_ffmpeg.cpp
@@ -508,7 +508,7 @@ void FfmpegEncoder::ScaleVideo()
 			}
 			frame->width               = ScaledWidth;
 			frame->height              = ScaledHeight;
-			frame->format              = OutputFormat;
+			frame->format              = static_cast<int>(OutputFormat);
 			frame->pts                 = image->pts;
 			// 0 means auto-align based on current CPU
 			constexpr int memory_alignment = 0;
